# swagger.yaml
swagger: '2.0'
info:
  title: gas
  version: 1.0.0
host: 'vvdev.ru:1334'
schemes:
  - http
basePath: /api
produces:
  - application/json
paths:
...
  /admins/mrg_orgs:
    put:
      tags:
      - MRG
      summary: Обновить инфу о организации
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            inn:
              type: string
            kpp:
              type: string
      responses:
        200:
          description: ok
    get:
      tags:
      - MRG
      summary: Получить инфу о организации
      responses:
        200:
          description: ok
  /admins/consumers/notifications:
    post:
      tags:
      - Consumer notification
      summary: Создать уведомление
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            type:
              type: string
            text:
              type: string
            cons_UIDs:
              type: array
              items:
                type: string
            userIds:
              type: array
              items:
                type: string
            allConsumers:
              type: boolean
      responses:
        200:
          description: ok
    get:
      tags:
      - Consumer notification
      summary: Список уведомлений для админа
      parameters:
      - in: query
        name: limit
        type: string
      - in: query
        name: page
        type: string
      responses:
        200:
          description: ok  
...
  /admins/consumers/notifications/{id}:
    get:
      tags:
      - Consumer notification
      summary: уведомление по id для админа
      parameters:
      - in: path
        type: string
        name: id
        required: true
      responses:
        200:
          description: ok 
    delete:
      tags:
      - Consumer notification
      summary: удалить уведомление по id для админа
      parameters:
      - in: path
        type: string
        name: id
        required: true
      responses:
        200:
          description: ok 
  /admins/consumers/notifications_by_user:
    get:
      tags:
      - Consumer notification
      summary: Список уведомлений для пользователя
      parameters:
      - in: query
        name: new_n
        type: boolean
      responses:
        200:
          description: ok 
  /admins/consumers/notifications_by_user/{id}:
    get:
      tags:
      - Consumer notification
      summary: уведомление по id для пользователя
      parameters:
      - in: path
        type: string
        name: id
        required: true
      responses:
        200:
          description: ok  
...
  /admins/login:
    post:
      tags:
      - Admin
      summary: Вход
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            login:
              type: string
            password:
              type: string
      responses:
        200:
          description: Токен
  /admins/users:
    get:
      tags:
      - Admin
      summary: Список пользователей
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            email:
              type: string
            name:
              type: string
            page:
              type: number
            limit:
              type: number
      responses:
        200:
          description: ок
  /admins/users/{id}:
    get:
      tags:
      - Admin
      summary: Пользователь по id
      parameters:
      - in: path
        name: id
        type: string
        required: true
      responses:
        200:
          description: ок 
  /admins/users/consumers/{id}:
    get:
      tags:
      - Admin
      summary: Потребитель по id
      parameters:
      - in: path
        name: id
        type: string
        required: true
      responses:
        200:
          description: ок 
  /admins/users/blocked:
    put:
      tags:
      - Admin
      summary: Пользователь по id
      parameters:
      - in: body
        name: body
        schema: 
          type: object
          properties:
            blocked: 
              type: boolean
            userId:
              type: string
            consumerId:
              type: string
            subLoginId:
              type: string
      responses:
        200:
          description: ок 
  /admins/users/sub_logins/{id}:
    put:
      tags:
      - Admin
      summary: Пользователь по id
      parameters:
      - in: path
        name: id
        required: true
        type: string
      - in: body
        name: body
        schema: 
          type: object
          properties:
            essences:
              type: object
              properties:
                consumers:
                  type: array
                  items:
                    type: string
                locations:
                  type: array
                  items:
                    type: string
                mets:
                  type: array
                  items:
                    type: string
            tabs:
              $ref: '#/definitions/Tabs'
            roleIds:
              type: array
              items:
                type: string
      responses:
        200:
          description: ок 
...
  /users/login:
    post:
      tags:
      - User
      summary: Вход
      parameters:
      - in: body
        name: body
        schema:
          type: object
          properties:
            login:
              type: string
            password:
              type: string
      responses:
        200:
          description: Токен
          schema:
            type: object
            properties:
              token:
                type: string
              _id:
                type: string
                description: id пользователя
              consumers:
                  type: array
                  items:
                    type: object
                    properties:
                      login:
                        type: string
                      cons_UID:
                        type: string
              cons_UIDs:
                type: array
                items:
                  type: string
        401:
          description: Пароль неправильный
        500:
          description: Ошибка сервера
  /users/{id}:
    delete:
      tags:
      - User
      summary: удалить пользователя(будут удалены все консамеры и сублогины относящиеся к пользователю)
      parameters:
      - in: path
        name: id
        type: string
        required: true
      responses:
        200:
          description: ok
...
  /users/sub_logins/{id}:
    get:
      tags:
      - User
      parameters:
      - in: path
        name: id
        type: string
        required: true
      responses:
        200:
          description: ok
        500:
          description: server error
    put:
      tags:
      - User
      parameters:
      - in: path
        name: id
        type: string
        required: true
      - in: body
        name: body
        schema:
          type: object
          properties:
            roleIds:
              type: array
              items:
                type: string
            roleId:
              type: string
            essences:
              type: object
              properties:
                consumers:
                  type: array
                  items:
                    type: string
                locations:
                  type: array
                  items:
                    type: string
                mets:
                  type: array
                  items:
                    type: string
            blocked:
              type: boolean
            firstName:
              type: string
            lastName:
              type: string
            middleName:
              type: string
            tabs:
              $ref: '#/definitions/Tabs'
            position:
              type: string
      responses:
        200:
          description: ok
    delete: 
      tags:
      - User
      summary: delete role
      parameters:
      - in: path
        name: id
        type: string
        required: true
      responses:
        200:
          description: ok
        404:
          description: not found
        500:
          description: server error
...
  /users/consumers/{id}:
    put:
      tags:
      - User
      summary: Изменить консамера
      parameters:
      - in: path
        name: id
        type: string
        required: true
      - in: body
        name: body
        schema:
          type: object
          properties:
            name:
              type: string
      responses:
        200:
          description: OK
...
  /contracts:
    get:
      tags:
      - Contract
      summary: Список договоров клиента
      responses:
        200:
          description: Список договоров
          schema:
            type: object
            properties:
              pagination:
                $ref: '#/definitions/Pagination'
              contracts:
                type: array
                items:
                  $ref: '#/definitions/ContractList'
        500:
          description: Ошибка сервера
  /contracts/{id}:
    get:
      tags:
      - Contract
      summary: Инфа по договору
      parameters:
      - in: path
        type: string
        name: id
        required: true
      responses:
        200:
          description: Инфа по договору
          schema:
            $ref: '#/definitions/Contract'
        500:
          description: Ошибка сервера
...
definitions:
...
  Pagination:
    type: object
    properties:
      next: 
        type: number
      prev:
        type: number
      total:
        type: number
      limit:
        type: number
      page:
        type: number
  Tabs:
    type: array
    items:
      $ref: '#/definitions/Tab'
  Tab:
    type: object
    properties:
      name:
        type: string
      read: 
        type: boolean
      write:
        type: boolean
...
  Contract:
    type: object
    properties:
      contr_num:
        type: string
      contr_date:
        type: string
      contr_status:
        type: string
      curr_balance:
        type: string
      contr_st_date:  
        type: string
      contr_end_date:
        type: string
      Abon_dep:
        type: string
      cur_fio:
        type: string
      cur_phone:
        type: string
      cur_email:
        type: string
      contr_scan:
        type: string
      FZ44:
        type: boolean
      FZ223:
        type: boolean
      Budg_ch_contr:
        type: string
      budg_level:
        type: string
      receivers:
        type: array
        items:
          type: object
          properties:
            example:
              type: string
      budg_limit:
        type: string
      budg_type:  
        type: string
      budg_contr_price:
        type: number
      other_contr_price:
        type: number
      budg_gas_vol:
        type: number
      other_gas_vol:
        type: number
      mets:
        type: array
        items:
          type: object
          properties:
            example:
              type: string
      Cons_points:
        type: array
        items:
          type: object
          properties:
            example:
              type: string
...
  ContractList:
    type: array
    items:
      type: object
      properties:
        contr_UID:
          type: string
        cons_UID:
          type: string
        contr_num:
          type: string
        contr_date:
          type: string
        contr_status:
          type: string
        curr_bakance:
          type: string
...